Use HEAD version of casacore-data
Compiling the python package does not depend on the exact version
of the measures data. The latest is fine.

This avoids crashing the build if casacore-data is out of date, which
happens from time to time. Since ftp.astron.nl only stores the latest
3 months or so, this requires casacore-data to be updated every 3 months
in sync with it, and it doesn't always happen.